refactor: fold turnkey tempo into turnkey adapter
