+++ playgrounds/web/src/provider.ts
@@ -11,13 +11,7 @@ import { Mppx } from 'mppx/client'
 import { generatePrivateKey } from 'viem/accounts'
 import { Account } from 'viem/tempo'
 
-export type AdapterType =
-  | 'secp256k1'
-  | 'turnkey'
-  | 'turnkeyTempo'
-  | 'webAuthn'
-  | 'tempoWallet'
-  | 'dialogRefImpl'
+export type AdapterType = 'secp256k1' | 'turnkey' | 'webAuthn' | 'tempoWallet' | 'dialogRefImpl'
 export type Env = 'mainnet' | 'testnet' | 'devnet'
 export type DialogMode = 'iframe' | 'popup'
 export type ProviderValue = ReturnType<typeof Provider.create>
@@ -101,7 +95,7 @@ export function createProvider(
     })
   }
 
-  if (adapterType === 'turnkey' || adapterType === 'turnkeyTempo')
+  if (adapterType === 'turnkey')
     return Provider.create({
       ...providerOptions,
       adapter: local({
@@ -157,7 +151,7 @@ export function switchProvider(next: ProviderValue) {
 export function switchDialogMode(mode: DialogMode, adapterType: AdapterType = 'tempoWallet') {
   dialogMode = mode
   Mppx.restore()
-  provider = createProvider(adapterType)
+  if (adapterType !== 'turnkey') provider = createProvider(adapterType)
 }
 
 export function switchTheme(
@@ -166,5 +160,5 @@ export function switchTheme(
 ) {
   theme = next
   Mppx.restore()
-  provider = createProvider(adapterType)
+  if (adapterType !== 'turnkey') provider = createProvider(adapterType)
 }
