+++ src/index.ts
@@ -13,7 +13,6 @@ export * as TrustedHosts from './core/TrustedHosts.js'
 export * as WebAuthnCeremony from './core/WebAuthnCeremony.js'
 export { dialog, dialog as tempoWallet } from './core/adapters/dialog.js'
 export { local } from './core/adapters/local.js'
-export { turnkeyTempo } from './core/adapters/turnkeyTempo.js'
 export { turnkey } from './core/adapters/turnkey.js'
 export { webAuthn } from './core/adapters/webAuthn.js'
 export { dangerous_secp256k1 } from './core/adapters/dangerous_secp256k1.js'
