+++ src/react/index.ts
@@ -1,2 +1,2 @@
 export * as Remote from './Remote.js'
-export { useTurnkeyAdapter, useTurnkeyTempoAdapter } from './Turnkey.js'
+export { useTurnkeyAdapter } from './Turnkey.js'
